Refactor

+++ MODELS/veicolo.cpp
@@ -2,7 +2,7 @@
 
 Veicolo::Veicolo(){}
 
-Veicolo::Veicolo(string targa, string posizione, double km, u_int numeroPosti, u_int ingombro):
+Veicolo::Veicolo(string targa, string posizione, double km, unsigned short numeroPosti, unsigned short ingombro):
     _targa(targa), _posizione(posizione), _chilometraggio(km), _capacitaPosti(numeroPosti), _ingombro(ingombro),
     _numeroUsi(0), _tempoServizio(0), _statoAttuale(libero), _inRiserva(false), _serveAssistenza(false) {}
